Support simultaneous animation of multiple objects
Use d3 selectAll() rather than select() as the underlying selection
mechanism. This makes it easy to ask for very inefficient animations,
but is useful in enough situations to be worth this risk.

I have bumped the version to 1.1 because this is potentially a
significant behaviour change, which could break existing applications.