Panel slider bug
There was a bug affecting the behaviour of the panel slider next
button under certain circumstances. The panel elements store their
index number as a DOM attribute; but DOM attributes are implicitly
strings, and we were failing to convert the index to a number before
setting the selected_panel member to it. Therefore adding 1 to the
selected panel, when the next button was pressed, would concatenate
'1' to the index rather than adding 1 to it.